Move uvcalibrate and apply_uvflag into utils subpackage
