+++ src/pyuvdata/utils/__init__.py
@@ -18,6 +18,7 @@ RADIAN_TOL = 1 * 2 * np.pi * 1e-3 / (60.0 * 60.0 * 360.0)
 LST_RAD_TOL = 2 * np.pi * 5e-3 / (86400.0)
 
 # these seem to be necessary for the installed package to access these submodules
+from . import apply_uvflag  # noqa
 from . import array_collapse  # noqa
 from . import bls  # noqa
 from . import bltaxis  # noqa
@@ -31,14 +32,17 @@ from . import pol  # noqa
 from . import redundancy  # noqa
 from . import times  # noqa
 from . import tools  # noqa
+from . import uvcalibrate  # noqa
 
 # Add things to the utils namespace used by outside packages
+from .apply_uvflag import apply_uvflag  # noqa
 from .array_collapse import collapse  # noqa
 from .bls import *  # noqa
 from .coordinates import *  # noqa
 from .phasing import uvw_track_generator  # noqa
 from .pol import *  # noqa
 from .times import get_lst_for_time  # noqa
+from .uvcalibrate import uvcalibrate  # noqa
 
 # deprecated imports
 
@@ -117,107 +121,3 @@ def _fits_indexhdus(hdulist):
     )
 
     return _indexhdus(hdulist)
-
-
-def uvcalibrate(uvdata, uvcal, **kwargs):
-    """
-    Calibrate a UVData object with a UVCal object.
-
-    Deprecated, use pyuvdata.uvcalibrate
-
-    Parameters
-    ----------
-    uvdata : UVData object
-        UVData object to calibrate.
-    uvcal : UVCal object
-        UVCal object containing the calibration.
-    inplace : bool, optional
-        if True edit uvdata in place, else return a calibrated copy
-    prop_flags : bool, optional
-        if True, propagate calibration flags to data flags
-        and doesn't use flagged gains. Otherwise, uses flagged gains and
-        does not propagate calibration flags to data flags.
-    Dterm_cal : bool, optional
-        Calibrate the off-diagonal terms in the Jones matrix if present
-        in uvcal. Default is False. Currently not implemented.
-    flip_gain_conj : bool, optional
-        This function uses the UVData ant_1_array and ant_2_array to specify the
-        antennas in the UVCal object. By default, the conjugation convention, which
-        follows the UVData convention (i.e. ant2 - ant1), is that the applied
-        gain = ant1_gain * conjugate(ant2_gain). If the other convention is required,
-        set flip_gain_conj=True.
-    delay_convention : str, optional
-        Exponent sign to use in conversion of 'delay' to 'gain' cal_type
-        if the input uvcal is not inherently 'gain' cal_type. Default to 'minus'.
-    undo : bool, optional
-        If True, undo the provided calibration. i.e. apply the calibration with
-        flipped gain_convention. Flag propagation rules apply the same.
-    time_check : bool
-        Option to check that times match between the UVCal and UVData
-        objects if UVCal has a single time or time range. Times are always
-        checked if UVCal has multiple times.
-    ant_check : bool
-        Option to check that all antennas with data on the UVData
-        object have calibration solutions in the UVCal object. If this option is
-        set to False, uvcalibrate will proceed without erroring and data for
-        antennas without calibrations will be flagged.
-
-    Returns
-    -------
-    UVData, optional
-        Returns if not inplace
-
-    """
-    from ..uvcalibrate import uvcalibrate
-
-    warnings.warn(
-        "uvcalibrate has moved, please import it as 'from pyuvdata import "
-        "uvcalibrate'. This warnings will become an error in version 3.2",
-        DeprecationWarning,
-    )
-
-    return uvcalibrate(uvdata, uvcal, **kwargs)
-
-
-def apply_uvflag(uvd, uvf, **kwargs):
-    """
-    Apply flags from a UVFlag to a UVData instantiation.
-
-    Deprecated, use pyuvdata.apply_uvflag
-
-    Note that if uvf.Nfreqs or uvf.Ntimes is 1, it will broadcast flags across
-    that axis.
-
-    Parameters
-    ----------
-    uvd : UVData object
-        UVData object to add flags to.
-    uvf : UVFlag object
-        A UVFlag object in flag mode.
-    inplace : bool
-        If True overwrite flags in uvd, otherwise return new object
-    unflag_first : bool
-        If True, completely unflag the UVData before applying flags.
-        Else, OR the inherent uvd flags with uvf flags.
-    flag_missing : bool
-        If input uvf is a baseline type and antpairs in uvd do not exist in uvf,
-        flag them in uvd. Otherwise leave them untouched.
-    force_pol : bool
-        If True, broadcast flags to all polarizations if they do not match.
-        Only works if uvf.Npols == 1.
-
-    Returns
-    -------
-    UVData
-        If not inplace, returns new UVData object with flags applied
-
-    """
-    from ..apply_uvflag import apply_uvflag
-
-    warnings.warn(
-        "uvcalibrate has moved, please import it as 'from pyuvdata import "
-        "uvcalibrate'. This warnings will become an error in version 3.2",
-        DeprecationWarning,
-    )
-
-    return apply_uvflag(uvd, uvf, **kwargs)
